ParseConfig/requirements.txt
PyYAML~=6.0
streamlit~=1.16.0
pandas~=1.5.2
xlrd
openpyxl